IncTA.jl

This project implements some Technical Analysis Indicators in Julia in an incremental approach.

It's inspired by Python project talipp which is used as "reference implementation" for unit tests.

It depends especially on OnlineStatsBase.jl and OnlineStats.jl.

🚧 This software is under construction. API can have breaking changes.

Package Features

Calculate new value of some technical analysis indicators when new incoming are received

Install

Open Julia command line interface.

Type ] dev https://github.com/femtotrader/IncTA.jl/

Function Documentation (alphabetically ordered)

IncTA.ADX — Type

ADX{Tohlcv,S}(; di_period = 14, adx_period = 14)

The ADX type implements an Average Directional Index indicator.

IncTA.ALMA — Type

ALMA{T}(; period = ALMA_PERIOD, offset = ALMA_OFFSET, sigma = ALMA_SIGMA)

The ALMA type implements an Arnaud Legoux Moving Average indicator.

IncTA.AO — Type

AO{Tohlcv,S}(; fast_period = AO_FAST_PERIOD, slow_period = AO_SLOW_PERIOD, fast_ma = SMA, slow_ma = SMA)

The AO type implements an Awesome Oscillator indicator.

IncTA.ATR — Type

ATR{Tohlcv,S}(; period = ATR_PERIOD)

The ATR type implements an Average True Range indicator.

IncTA.AccuDist — Type

AccuDist{Tohlcv,S}()

The AccuDist type implements an Accumulation and Distribution indicator.

IncTA.Aroon — Type

Aroon{Tohlcv,S}(; period = Aroon_PERIOD)

The Aroon type implements an Aroon indicator.

IncTA.BB — Type

BB{T}(; period = BB_PERIOD, std_dev_multiplier = BB_STD_DEV_MULTIPLIER, ma = SMA)

The BB type implements Bollinger Bands indicator.

IncTA.BOP — Type

BOP{Tohlcv,S}()

The BOP type implements a Balance Of Power indicator.

IncTA.CCI — Type

CCI{Tohlcv,S}(; period=CCI_PERIOD)

The CCI type implements a Commodity Channel Index.

IncTA.CHOP — Type

CHOP{Tohlcv,S}(; period = CHOP_PERIOD)

The CHOP type implements a Choppiness Index indicator.

IncTA.ChaikinOsc — Type

ChaikinOsc{Tohlcv,S}(; fast_period = ChaikinOsc_FAST_PERIOD, slow_period = ChaikinOsc_SLOW_PERIOD, fast_ma = EMA, slow_ma = EMA)

The ChaikinOsc type implements a Chaikin Oscillator.

IncTA.ChandeKrollStop — Type

ChandeKrollStop{Tohlcv,S}(; atr_period = ChandeKrollStop_ATR_PERIOD, atr_mult = ChandeKrollStop_ATR_MULT, period = ChandeKrollStop_PERIOD)

The ChandeKrollStop type implements a ChandeKrollStop indicator.

IncTA.CoppockCurve — Type

CoppockCurve{T}(; fast_roc_period = CoppockCurve_FAST_ROC_PERIOD, slow_roc_period = CoppockCurve_SLOW_ROC_PERIOD, wma_period = CoppockCurve_WMA_PERIOD)

The CoppockCurve type implements a Coppock Curve indicator.

IncTA.DEMA — Type

DEMA{T}(; period = DEMA_PERIOD)

The DEMA type implements a Double Exponential Moving Average indicator.

IncTA.DPO — Type

DPO{T}(; period = DPO_PERIOD)

The DPO type implements a Detrended Price Oscillator indicator.

IncTA.DonchianChannels — Type

DonchianChannels{Tohlcv,S}(; period = DonchianChannels_ATR_PERIOD)

The DonchianChannels type implements a Donchian Channels indicator.

IncTA.EMA — Type

EMA{T}(; period=EMA_PERIOD)

The EMA type implements an Exponential Moving Average indicator.

IncTA.EMV — Type

EMV{Tohlcv,S}(; period = EMV_PERIOD, volume_div = EMV_VOLUME_DIV, ma = SMA)

The EMV type implements a Ease of Movement indicator.

IncTA.ForceIndex — Type

ForceIndex{Tohlcv,S}(; period = ForceIndex_PERIOD, ma = EMA)

The ForceIndex type implements a Force Index indicator.

IncTA.HMA — Type

HMA{T}(; period = HMA_PERIOD)

The HMA type implements a Hull Moving Average indicator.

IncTA.KAMA — Type

KAMA{T}(; period = KAMA_PERIOD)

The KAMA type implements a Kaufman's Adaptive Moving Average indicator.

IncTA.KST — Type

KST{T}(;
    roc1_period = KST_ROC1_PERIOD,
    roc1_ma_period = KST_ROC1_MA_PERIOD,
    roc2_period = KST_ROC2_PERIOD,
    roc2_ma_period = KST_ROC2_MA_PERIOD,
    roc3_period = KST_ROC3_PERIOD,
    roc3_ma_period = KST_ROC3_MA_PERIOD,
    roc4_period = KST_ROC4_PERIOD,
    roc4_ma_period = KST_ROC4_MA_PERIOD,
    signal_period = KST_SIGNAL_PERIOD,
    ma = SMA
)

The KST type implements Know Sure Thing indicator.

IncTA.KVO — Type

KVO{Tohlcv,S}(; fast_period = KVO_FAST_PERIOD, slow_period = KVO_SLOW_PERIOD, ma = EMA)

The KVO type implements a Klinger Volume Oscillator.

IncTA.KeltnerChannels — Type

KeltnerChannels{Tohlcv,S}(; ma_period = KeltnerChannels_MA_PERIOD, atr_period = KeltnerChannels_ATR_PERIOD, atr_mult_up = KeltnerChannels_ATR_MULT_UP, atr_mult_down = KeltnerChannels_ATR_MULT_DOWN, ma = EMA)

The KeltnerChannels type implements a Keltner Channels indicator.

IncTA.MACD — Type

MACD{T}(; fast_period = MACD_FAST_PERIOD, slow_period = MACD_SLOW_PERIOD, signal_period = MACD_SIGNAL_PERIOD, ma = EMA)

The MACD type implements Moving Average Convergence Divergence indicator.

IncTA.MassIndex — Type

MassIndex{T,S}(; ma_period = MassIndex_MA_PERIOD, ma_ma_period = MassIndex_MA_MA_PERIOD, ma_ratio_period = MassIndex_MA_RATIO_PERIOD, ma = EMA)

The MassIndex type implements a Commodity Channel Index.

IncTA.McGinleyDynamic — Type

McGinleyDynamic{T}(; period = McGinleyDynamic_PERIOD)

The McGinleyDynamic type implements a McGinley Dynamic indicator.

IncTA.MeanDev — Type

MeanDev{T}(; period = MeanDev_PERIOD, ma = SMA)

The MeanDev type implements a Mean Deviation indicator.

IncTA.OBV — Type

OBV{Tohlcv,S}()

The OBV type implements On Balance Volume indicator.

IncTA.ParabolicSAR — Type

ParabolicSAR{Tohlcv,S}(; atr_period = ParabolicSAR_ATR_PERIOD, mult = ParabolicSAR_MULTIPLIER)

The ParabolicSAR type implements a Super Trend indicator.

IncTA.ROC — Type

ROC{T}(; period = ROC_PERIOD)

The ROC type implements a Rate Of Change indicator.

IncTA.RSI — Type

RSI{T}(; period = SMA_PERIOD)

The RSI type implements a Relative Strength Index indicator.

IncTA.SMA — Type

SMA{T}(; period = SMA_PERIOD)

The SMA type implements a Simple Moving Average indicator.

IncTA.SMMA — Type

SMMA{T}(; period = SMA_PERIOD)

The SMMA type implements a SMoothed Moving Average indicator.

IncTA.SOBV — Type

SOBV{Tohlcv,S}(; period = SOBV_PERIOD, ma = SMA)

The SOBV type implements a Smoothed On Balance Volume indicator.

IncTA.StdDev — Type

StdDev{T}(; period = StdDev_PERIOD)

The StdDev type implements a Standard Deviation indicator.

IncTA.Stoch — Type

Stoch{Tohlcv,S}(; period = STOCH_PERIOD, smoothing_period = STOCH_SMOOTHING_PERIOD, ma = SMA)

The Stoch type implements the Stochastic indicator.

IncTA.SuperTrend — Type

SuperTrend{Tohlcv,S}(; atr_period = SuperTrend_ATR_PERIOD, mult = SuperTrend_MULTIPLIER)

The SuperTrend type implements a Super Trend indicator.

IncTA.TEMA — Type

TEMA{T}(; period = TEMA_PERIOD, ma = EMA)

The TEMA type implements a Triple Exponential Moving Average indicator.

IncTA.UO — Type

UO{Tohlcv,S}(; fast_period = UO_FAST_PERIOD, mid_period = UO_MID_PERIOD, slow_period = UO_SLOW_PERIOD)

The UO type implements an Ultimate Oscillator.

IncTA.VTX — Type

VTC{Tohlcv,S}(; period = VTX_PERIOD)

The VTX type implements a Vortex Indicator.

IncTA.VWAP — Type

VWAP{Tohlcv,S}()

The VWAP type implements a Volume Weighted Moving Average indicator.

IncTA.VWMA — Type

VWMA{Tohlcv,S}(; period = VWMA_PERIOD)

The VWMA type implements a Volume Weighted Moving Average indicator.

IncTA.WMA — Type

WMA{T}(; period = WMA_PERIOD)

The WMA type implements a Weighted Moving Average indicator.